Modify used namespace in selected classes
The code was rearranged. Some of the classes that are used in the tests were in separate folders and namespaces were generated for them. Then folders were remove, class files were placed in the main directory of the project and after that the namespaces used in those files had to be changed