Instructions - add triggers and changed logic for the foreign keys
Added triggers on delete and so the cascade operations to work properly i changed the logic in some of the tables. The main problem was that because in FLIGHT we have two foreign keys referencing to the same other table, when deleting records from that other table it causes problems.